오류 수정 (login exterid long으로 변경)

+++ app/src/main/java/com/example/grandpa/SignupWithKakaoActivity.kt
@@ -79,7 +79,7 @@ class SignupWithKakaoActivity:AppCompatActivity() {
             }
 
             //db에 저장
-            InfoData.putLong("externalId", kakaoInfo.externalId.toLong())
+            InfoData.putLong("externalId", kakaoInfo.externalId)
             InfoData.putString("loginType", "kakao")
             InfoData.putString("name", kakaoInfo.nickname)
